separate out staging manager
